+++ SourceryRuntime/Sources/AST/Method.swift
@@ -281,43 +281,43 @@ extension Array where Element == ClosureParameter {
     // sourcery: skipEquality, skipDescription, skipCoding, skipJSExport
     /// :nodoc:
     @available(*, deprecated, message: "Use isConvenienceInitializer instead") public var isConvenienceInitialiser: Bool {
-        return attributes[Attribute.Identifier.convenience.name] != nil
+        isConvenienceInitialiser
     }
 
     // sourcery: skipEquality, skipDescription
     /// Whether method is a convenience initializer
     public var isConvenienceInitializer: Bool {
-        return attributes[Attribute.Identifier.convenience.name] != nil
+        modifiers.contains { $0.name == "convenience" }
     }
 
     // sourcery: skipEquality, skipDescription
     /// Whether method is required
     public var isRequired: Bool {
-        return attributes[Attribute.Identifier.required.name] != nil
+        modifiers.contains { $0.name == "required" }
     }
 
     // sourcery: skipEquality, skipDescription
     /// Whether method is final
     public var isFinal: Bool {
-        return attributes[Attribute.Identifier.final.name] != nil
+        modifiers.contains { $0.name == "final" }
     }
 
     // sourcery: skipEquality, skipDescription
     /// Whether method is mutating
     public var isMutating: Bool {
-        return attributes[Attribute.Identifier.mutating.name] != nil
+        modifiers.contains { $0.name == "mutating" }
     }
 
     // sourcery: skipEquality, skipDescription
     /// Whether method is generic
     public var isGeneric: Bool {
-        return shortName.hasSuffix(">")
+        shortName.hasSuffix(">")
     }
 
     // sourcery: skipEquality, skipDescription
     /// Whether method is optional (in an Objective-C protocol)
     public var isOptional: Bool {
-        return attributes[Attribute.Identifier.optional.name] != nil
+        modifiers.contains { $0.name == "optional" }
     }
 
     /// Annotations, that were created with // sourcery: annotation1, other = "annotation value", alterantive = 2